requirements.txt
python-docx
google-generativeai
tiktoken
pydantic
python-dotenv
